rename footer to be lowercase
