+++ packages/dashboard-site/src/components/footer/index.tsx
